feat: 소셜 로그인 기능 수정 및 dockerFile 추가
