+++ Dockerfile
@@ -1,0 +1,7 @@
+FROM openjdk:17-jdk-alpine
+
+ARG JAR_FILE=build/libs/*.jar
+
+COPY ${JAR_FILE} app.jar
+
+ENTRYPOINT ["java","-jar","/app.jar"]
