+++ src/main/java/gdg/pium/global/auth/controller/OAuthController.java
@@ -1,16 +1,15 @@
 package gdg.pium.global.auth.controller;
 
-import gdg.pium.global.auth.dto.response.KakaoLoginResponseDto;
-import gdg.pium.global.auth.dto.response.KakaoLoginUrlResponseDto;
+import gdg.pium.global.auth.dto.request.KakaoLoginRequest;
+import gdg.pium.global.auth.dto.response.KakaoLoginResponse;
+import gdg.pium.global.auth.dto.response.KakaoLoginUrlResponse;
 import gdg.pium.global.auth.service.OAuthService;
+import gdg.pium.global.common.dto.ResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("api/v1/oauth")
@@ -19,30 +18,32 @@ public class OAuthController {
     private final OAuthService oAuthService;
 
     @Operation(
-        summary = "[공통] 카카오 로그인",
+        summary = "[공통] 카카오 로그인 URL 반환",
         description = "카카오 로그인 화면 URL을 반환합니다. 해당 URL로 이동해주세요.",
         responses = {
             @ApiResponse(responseCode = "200", description = "이동 후 사용자가 카카오 로그인을 마칠 시 카카오 인증 서버는 /kakao/callback으로 redirect 시킵니다."),
         }
     )
-    @GetMapping("/kakao")
-    public ResponseEntity<KakaoLoginUrlResponseDto> kakaoLogin() {
-        return ResponseEntity.ok(
-            KakaoLoginUrlResponseDto.builder()
+    @GetMapping("/kakao/auth-uri")
+    public ResponseDto<KakaoLoginUrlResponse> kakaoLogin() {
+        return ResponseDto.ok(
+            KakaoLoginUrlResponse.builder()
             .authorizationUrl(oAuthService.getKakaoAuthorizationUrl())
             .build()
         );
     }
 
     @Operation(
-        summary = "[공통] 카카오 서버가 사용자로 하여금 redirect 시키는 API endpoint",
-        description = "서버의 redirect 응답을 그대로 수행하면, 해당 API로 인가 코드와 함께 요청을 보내고, 서버는 카카오 계정 정보를 이용해 로그인 또는 회원가입 절차(id, pw 저장)를 처리해 결과를 반환합니다.",
+        summary = "[공통] 카카오 로그인 후 사용자 등록 및 jwt 발급",
+        description = "accessToken 을 받아 카카오 서버로 부터 사용자 정보를 받고 그 정보를 DB에 등록한 뒤 jwt 발급",
         responses = {
             @ApiResponse(responseCode = "200", description = "isNewUser는 신규 회원인지의 여부. 신규 회원이라면 음식점|후원자|결식아동 중 하나를 선택하라는 화면으로 이동."),
         }
     )
-    @GetMapping("/kakao/callback")
-    public ResponseEntity<KakaoLoginResponseDto> kakaoCallback(@RequestParam("code") String code) {
-        return ResponseEntity.ok(oAuthService.processKakaoCallback(code));
+    @GetMapping("/kakao/login")
+    public ResponseDto<KakaoLoginResponse> kakaoCallback(
+            @RequestBody KakaoLoginRequest request
+    ) {
+        return ResponseDto.ok(oAuthService.processKakaoLogin(request));
     }
 }
