+++ src/main/java/gdg/pium/global/auth/dto/request/AccountLoginRequest.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
-public record AccountLoginRequestDto(
+public record AccountLoginRequest(
     @NotBlank(message = "이메일은 필수 항목입니다")
     @Pattern(
         regexp = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
