+++ src/main/java/gdg/pium/global/auth/dto/request/KakaoLoginRequest.java
@@ -1,0 +1,8 @@
+package gdg.pium.global.auth.dto.request;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record KakaoLoginRequest(
+        @JsonProperty("access_token") String accessToken
+) {
+}
