+++ src/main/java/gdg/pium/global/auth/dto/request/TokenRefreshRequest.java
@@ -2,7 +2,7 @@ package gdg.pium.global.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record TokenRefreshRequestDto(
+public record TokenRefreshRequest(
     @NotBlank(message = "refreshToken은 필수 항목입니다")
     String refreshToken
 ) {
